refactor/#470 : FoodEntity를 domain 패키지로 이동
