+++ BE/exceed/src/main/java/com/gaebaljip/exceed/common/dto/PageableFoodDTO.java
@@ -2,6 +2,6 @@ package com.gaebaljip.exceed.common.dto;
 
 import java.util.List;
 
-import com.gaebaljip.exceed.adapter.out.jpa.food.FoodEntity;
+import com.gaebaljip.exceed.application.domain.food.FoodEntity;
 
 public record PageableFoodDTO(List<FoodEntity> foodEntities, Boolean hasNext, int size) {}
